Configuration form seperated out and reused across Patterns and Applications
